requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
matplotlib
google-generativeai
google-genai
requests
folium